pyproject.toml
[tool.poetry]
name = "atomic-rag-system"
version = "0.1.0"
description = "Multi-agent RAG system for PDF product search using SQLite + Qdrant"
authors = ["Konstantin Kuteykin-Teplyakov <konstantin.neurocreate@gmail.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.10"

# Core framework - simplified without atomic-agents dependency
# atomic-agents = ">=0.1.0,<2.0.0"  # Commented out - using custom base classes

# LLM & API clients
mistralai = ">=0.4.0,<2.0.0"
# openai = ">=1.0.0,<2.0.0"  # Optional: for embeddings - not needed

# Vector & Storage
qdrant-client = ">=1.6.0,<2.0.0"

# Embeddings & ML - CPU-only, German-friendly models
sentence-transformers = ">=2.2.0,<3.0.0"
# Using CPU-only transformers with German models
transformers = ">=4.30.0,<5.0.0"
tokenizers = ">=0.13.0,<1.0.0"
# CPU-only numpy for embeddings
numpy = ">=1.21.0,<2.0.0"
scikit-learn = ">=1.3.0,<2.0.0"

# Data processing
pydantic = ">=2.0.0,<3.0.0"
python-dotenv = ">=1.0.0,<2.0.0"
pydantic-settings = ">=2.0.0,<3.0.0"

# PDF processing
pypdf2 = ">=3.0.0,<4.0.0"
pillow = ">=9.0.0,<11.0.0"

# Database & search
# aiosqlite = ">=0.19.0,<1.0.0"  # Not needed - using standard sqlite3

# CLI & utilities
rich = ">=13.0.0,<14.0.0"  # Beautiful terminal output
typer = ">=0.9.0,<1.0.0"  # CLI framework
loguru = ">=0.7.0,<1.0.0"  # Better logging

# HTTP
requests = ">=2.31.0,<3.0.0"
httpx = ">=0.25.0,<1.0.0"

[tool.poetry.group.dev.dependencies]
# Testing
pytest = ">=7.4.0,<8.0.0"
pytest-asyncio = ">=0.21.0,<1.0.0"
pytest-cov = ">=4.1.0,<5.0.0"
pytest-mock = ">=3.12.0,<4.0.0"

# Code quality
black = ">=23.0.0,<24.0.0"
isort = ">=5.12.0,<6.0.0"
flake8 = ">=6.1.0,<7.0.0"
mypy = ">=1.7.0,<2.0.0"
pylint = ">=3.0.0,<4.0.0"

# Development tools
ipython = ">=8.12.0,<9.0.0"
jupyter = ">=1.0.0,<2.0.0"

[tool.poetry.scripts]
# Main CLI entry point
atomic-rag = "main:app"

# Database initialization
init-db = "scripts.init_db:main"

# Test scripts
test-basic = "test_basic:main"
test-ocr = "test_ocr_models:main"
test-translation = "test_translation:main"

[tool.black]
line-length = 100
target-version = ['py310']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=src --cov-report=html --cov-report=term"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"